make footer links easier to click using padding
Now, clicking the area around the links will trigger the nearest link. This makes the buttons easier to click, and less annoying.

These styles are still not perfect. In Firefox, the whitespace in the HTML adds unclickable space between the links. And there is some extra unclickable space below the links, too. You would probably need to use floats, along with the clearfix, to fix these problems. That’s too big a change for me to make in the GitHub editor. But at least these styles are better than before – the footer looks very similar, but the links are easier to click.